+++ tests/acceptance/config/behat.yml
@@ -11,6 +11,7 @@ default:
 
         - AppNavigationContext
         - FeatureContext
+        - FileListContext
         - FilesAppContext
         - FilesSharingAppContext
         - LoginPageContext
